requirements.txt
pyarrow
pandas
google-cloud-bigquery
google-cloud-storage
torch
scipy
numpy
matplotlib
flask
flask-assets
flask-caching
flask-restful
utm
python-dotenv
timezonefinder
cssmin
jsmin
flake8
gunicorn
